backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.2.1
